requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
-r requirements/production.txtDjango==1.5.1
GitPython==0.3.2.RC1
Jinja2==2.6
PIL==1.1.7
PyYAML==3.10
Pygments==1.6
South==0.7.6
Sphinx==1.2b1
amqp==1.0.11
anyjson==0.3.3
async==0.6.1
beautifulsoup4==4.1.3
behave==1.2.2
billiard==2.7.3.28
bpython==0.12
celery==3.0.19
coverage==3.6
django-admin-bootstrapped==0.3.2
django-behave==0.0.5
django-braces==1.0.0
django-celery==3.0.17
django-debug-toolbar==0.9.4
django-discover-runner==0.3
django-dynamic-settings==0.1
django-git==0.1
django-model-utils==1.3.1
docutils==0.10
ez-setup==0.9
fuzzywuzzy==0.1
gherkin==3
gitdb==0.5.4
gunicorn==0.17.0
html5lib==0.95
icalendar==3.1
ipdb==0.7
ipython==0.13.1
kombu==2.5.10
lettuce==0.2.14
logutils==0.3.3
lxml==3.1.1
markdown2==2.1.0
mechanize==0.2.5
paramiko==1.10.1
parse==1.6.1
psycopg2==2.5
pyPdf==1.13
pycrypto==2.6
pyparsing==1.5.7
python-dateutil==2.1
pytz==2012f
reportlab==2.5
selenium==2.31.0
simplejson==2.6.2
six==1.3.0
smmap==0.8.2
splinter==0.5.0
sure==1.1.7
wsgi-intercept==0.5.1
wsgiref==0.1.2
xhtml2pdf==0.0.4
